Add Snowflake ids to BaseModel, fix model_id length in history_of_operations table

+++ database/migrations/2021_03_02_174507_create_history_of_operations_table.php
@@ -17,7 +17,7 @@ class CreateHistoryOfOperationsTable extends Migration
             $table->id();
             $table->unsignedTinyInteger('operation_type');
             $table->unsignedTinyInteger('model_type');
-            $table->unsignedInteger('model_id');
+            $table->unsignedBigInteger('model_id');
             $table->jsonb('model');
             $table->timestamps();
         });
